# Conflicts:
#	Assets/Scenes/Map1RE.unity
#	UserSettings/Layouts/default-2021.dwlt

+++ Assets/Member/tei/Boss/Boss_Bowman.cs
@@ -25,7 +25,7 @@ public class Boss_Bowman : MonoBehaviour
     //クールタイム用
     private float Boss_Cool_time = 0;
     [SerializeField, Header("攻撃クールタイム")]
-    private int Boss_Atk_time = 2;
+    private float Boss_Atk_time = 2;
 
     //攻撃
     private int Boss_Archer_Atk1;
